pyproject.toml
[tool.poetry]
name = "presentation-rag"
version = "0.1.0"
description = "Presentation RAG Raft Internship"
authors = ["Ilia Tambovtsev <ilia.tambovtsev@raftds.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.5"
notebook = "^7.2.2"
nbclassic = "^1.1.0"
jupyter-console = "^6.6.3"
jupyter = "^1.1.1"
pdf2image = "^1.17.0"
langchain = "^0.3.4"
langchain-community = "^0.3.3"
langchain-openai = "^0.2.3"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
chromadb = "^0.5.20"
gradio = "^5.6.0"
gradio-pdf = "^0.0.19"
tabulate = "^0.9.0"
seaborn = "^0.13.2"
tiktoken = "^0.8.0"
langsmith = "^0.1.147"
mlflow = "^2.18.0"
fire = "^0.7.0"
nltk = "^3.9.1"
gspread-pandas = "^3.3.0"
pymupdf = "^1.25.1"
llama-parse = "^0.5.18"
pillow = "^11.0.0"
dash = "^2.18.2"
plotly = "^5.24.1"
dash-bootstrap-components = "^1.6.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
